fixed errors in test code
